requirements.txt
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langsmith==0.3.45
python-dotenv==1.1.0
openai==1.88.0
transformers==4.52.4
torch==2.7.1
sentence-transformers==4.1.0
pydantic==2.11.7
pydantic-settings==2.9.1
faiss-cpu==1.11.0
tiktoken==0.9.0
urllib3==2.4.0
requests==2.32.4
PyYAML==6.0.2
